requirements.txt
feedparser
requests
python-dotenv
python-dateutil
google-generativeai
